Route billing through the action()/actions convention; drop scattered try/catch
Follows the links endpoints' shape: logic lives in actions/billing.ts,
expected failures are returned as jsonapi payloads, and everything else
(auth throws, Stripe API errors) is converted by action()'s handleError
(E_UNAUTHORIZED_ACCESS -> 401, anything else -> jsonapi 500). The shared
accountForRequest prologue replaces the per-endpoint auth/account/admin
boilerplate.

The two remaining try/catch blocks are semantic, not error plumbing:
the /billing/success browser redirect must never render an error, and
the webhook controller needs 400-vs-500 distinction for Stripe retries.
